move fetching util
